clean up watchers
