+++ src/watcher/models/watcher/index.js
@@ -1,0 +1,7 @@
+import watcherPolymorphs from './polymorphs';
+import Watcher from './base';
+
+export {
+  Watcher,
+  watcherPolymorphs as watcherTypes,
+};
