+++ src/watcher/models/watcher/polymorphs/index.js
@@ -1,0 +1,9 @@
+import HttpWatcher from './http_watcher';
+import CloudwatchAlarmWatcher from './cloudwatch_alarm_watcher';
+import DummyWatcher from './dummy_watcher';
+
+export default {
+  HttpWatcher,
+  CloudwatchAlarmWatcher,
+  DummyWatcher,
+};
